third_party/blink/web_tests/flag-specific/disable-site-isolation-trials/virtual/soft-navigation-fp-fcp/external/wpt/soft-navigation-heuristics/navigation-api.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Test soft navigation with the Navigation API
  assert_greater_than: Soft navigation should have triggered at least an LCP entry expected a number greater than 1 but got 1
Harness: the test ran to completion.
